Move specs out of domain
