+++ module/spiff-element-units-rs/src/element_units.rs
@@ -3,8 +3,9 @@ use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 use std::error::Error;
 
-use crate::domain::{ElementIntrospection, IndexedVec, Map, WorkflowSpec};
+use crate::domain::{ElementIntrospection, IndexedVec, Map};
 use crate::reader;
+use crate::specs::{WorkflowSpec};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub enum ElementUnit {
